backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.4.0
packaging==24.0
platformdirs==4.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.0
virtualenv==20.26.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.0
